Moved vocabulary package from spi to core
Moved the vocabulary package from spi to core.  Modified all impacted import statements.  Modified pom.xml files to reflect the change.

Note:  Began having problems (zero arguments returned) with the extractTwoArguments test in VerifyCtlExecutor.java.  Traced the problem to the runOpts.addParam(kvp) line in the extractTestRunArguments() operation in CtlExecutor.java.  After debugging, the problem disappeared without any code changes.
