+++ teamengine-core/src/main/java/com/occamlab/te/vocabulary/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Standard vocabulary definitions.
+ *
+ * March 6, 2018 - Moved from spi to core by C. Heazel (WiSC)
+ *
+ */
+package com.occamlab.te.vocabulary;
