--- teamengine-spi/src/main/java/com/occamlab/te/spi/vocabulary/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Standard vocabulary definitions.
- */
-package com.occamlab.te.spi.vocabulary;
